requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
idna==2.9
pathlib==1.0.1
Pillow==7.1.2
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
